use MakeMove

+++ src/redux/engineMiddleware.ts
@@ -3,7 +3,7 @@ import { restartWorker, worker, write } from '../worker';
 import { RootState } from '../store';
 import { DEFAULT_POSITION } from 'chess.js';
 import { UCI_ENGINE_ERROR, TOGGLE_ENGINE, SET_HASH, SET_THREADS, SET_LINES, EngineAction } from './engineSlice';
-import { GameAction, GOTO_MOVE, GOTO_PATH, MAKE_MOVE } from './gameSlice';
+import { GameAction, GOTO_MOVE, GOTO_PATH, MakeGameMove } from './gameSlice';
 import { SET_SOURCE, TreeAction } from './treeSlice';
 
 export const setOption = (name: string, value: string | number) =>
@@ -25,44 +25,44 @@ export const createEngineMiddleware = (): Middleware => {
 
   return (store) => (next) => (action) => {
     const state = store.getState();
-    const workerAction = action as WorkerAction;
 
-    switch (workerAction.type) {
-      case UCI_ENGINE_ERROR.type:
-        if (workerAction.payload.includes('OOM')) {
-          worker.terminate();
-          restartWorker({ ...state.engine, threads: state.engine.threads / 2 });
+    if (action) {
+        const workerAction = action as WorkerAction;
+        switch (workerAction.type) {
+        case UCI_ENGINE_ERROR.type:
+            if (workerAction.payload.includes('OOM')) {
+            worker.terminate();
+            restartWorker({ ...state.engine, threads: state.engine.threads / 2 });
+            }
+            break;
+        case TOGGLE_ENGINE.type:
+            write(state.engine.running ? 'stop' : 'go infinite');
+            break;
+        case SET_HASH.type:
+            updateEngine(setOption('Hash', workerAction.payload), state);
+            break;
+        case SET_THREADS.type:
+            updateEngine(setOption('Threads', workerAction.payload), state);
+            break;
+        case SET_LINES.type:
+            updateEngine(setOption('MultiPV', workerAction.payload), state);
+            break;
+        case MakeGameMove.type:
+            updateEngine(setPos(workerAction.payload.after), state);
+            break;
+        case GOTO_MOVE.type:
+            updateEngine(setPos(workerAction.payload.fen), state);
+            break;
+        case GOTO_PATH.type:
+            updateEngine(setPos(workerAction.payload.at(-1)?.after || DEFAULT_POSITION), state);
+            break;
+        case SET_SOURCE.type:
+            updateEngine(setPos(DEFAULT_POSITION), state);
+            break;
+        default:
+            break;
         }
-        break;
-      case TOGGLE_ENGINE.type:
-        write(state.engine.running ? 'stop' : 'go infinite');
-        break;
-      case SET_HASH.type:
-        updateEngine(setOption('Hash', workerAction.payload), state);
-        break;
-      case SET_THREADS.type:
-        updateEngine(setOption('Threads', workerAction.payload), state);
-        break;
-      case SET_LINES.type:
-        updateEngine(setOption('MultiPV', workerAction.payload), state);
-        break;
-      case MAKE_MOVE.type:
-        updateEngine(setPos(workerAction.payload.after), state);
-        break;
-      case GOTO_MOVE.type:
-        updateEngine(setPos(workerAction.payload.fen), state);
-        break;
-      case GOTO_PATH.type:
-        updateEngine(setPos(workerAction.payload.at(-1)?.after || DEFAULT_POSITION), state);
-        break;
-      case SET_SOURCE.type:
-        updateEngine(setPos(DEFAULT_POSITION), state);
-        break;
-      default:
-        break;
+        return next(action);
     }
-
-
-    return next(action);
   };
 };
